feat(search): relocate SearchModal to components/ui/modal and update imports
